requirements.txt
torch
ctc_decoder
pandas
nltk
pyyaml
symspellpy
torchvision
natsort
pillow
dataclasses_json